+++ agents-manage-ui/cypress/e2e/components.cy.ts
@@ -4,7 +4,7 @@ import { randomId } from '../support/utils';
 
 describe('Components', () => {
   it('should create a new component when adding JSON schema properties with form builder and without `required` field', () => {
-    cy.visit('/default/projects/my-weather-project/components/new');
+    cy.visit('/default/projects/activities-planner/components/new');
     cy.get('input[name=name]').type(`test ${randomId()}`);
     cy.get('textarea[name=description]').type('test description');
     cy.contains('Add property').click();
@@ -13,25 +13,25 @@ describe('Components', () => {
     cy.contains('Save').click();
     cy.get('[data-sonner-toast]').contains('Component created').should('exist');
     // Should redirect
-    cy.location('pathname').should('eq', '/default/projects/my-weather-project/components');
+    cy.location('pathname').should('eq', '/default/projects/activities-planner/components');
   });
 
   it('should not override json schema when json mode is enabled by default', () => {
-    cy.visit('/default/projects/my-weather-project/components/weather-forecast');
+    cy.visit('/default/projects/activities-planner/components/activities');
     cy.get('[role=switch]').click();
     cy.reload();
-    cy.contains('Weather code at given time').should('exist');
+    cy.contains('The type of event').should('exist');
   });
 
   describe('inPreview', () => {
     it('should not have `inPreview` flag', () => {
-      cy.visit('/default/projects/my-weather-project/components/new');
+      cy.visit('/default/projects/activities-planner/components/new');
       cy.contains('Add property').should('exist');
       cy.contains('In Preview').should('not.exist');
     });
 
     it('should remove `inPreview` fields from editor', () => {
-      cy.visit('/default/projects/my-weather-project/components/new');
+      cy.visit('/default/projects/activities-planner/components/new');
       cy.get('[role=switch]').click();
 
       const editorValue = {